+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - `title` (`mj-button`), `sizes`/directional border attributes (`mj-image`), `role` (`mj-table`) and `font-weight` (`mj-accordion-title`)
   - `mj-carousel`: `thumbnails="supported"`
   - inherit `font-family` from `mj-accordion-element`
+  - `width="auto"` for `mj-table`
 - support `mj-include` with `type="css"` and `type="html"`
 - support `mj-include` of files which contain an `<mj-head>`
 - support `<mj-class>` with child elements (default attributes for nested elements)